+++ R/search_for_newer_data.R
@@ -4,53 +4,92 @@
 #' 
 #' Search in some folders if a TrialMaster database more recent than the current extration is present. By default, it will search the "data" folder and the OS usual "Downloads" folder. If a newer database is found, user will be asked if they want to move it to the "data" folder.
 #'
-#' @param project_name the project name, treated as a regex
-#' @param path the path vector to be searched, default to both "data" and the usual "Downloads" folder
+#' @param project archive path, giving the project name
+#' @param source the path vector to be searched, default to both "data" and the usual "Downloads" folder
+#' @param target the path where files should be copied
 #' @param ask whether to ask the user to move the file to "data"
 #' @param advice whether to advice how to move it, if `ask==FALSE`
 #'
 #' @return the path to the newer file, invisibly.
 #' @export
 #' @importFrom cli cli_inform
-#' @importFrom fs dir_ls file_copy path_file path_home
+#' @importFrom fs dir_ls file_copy path path_file path_home
 #' @importFrom glue glue
 #' @importFrom stringr str_detect
-search_for_newer_data = function(project_name, path=c("data", path_home("Downloads")), 
-                               ask=TRUE, advice=NULL){
-  files = dir_ls(path, type="file", regexp=glue(".*{project_name}.*\\.zip"))
+#' 
+#' @examples
+#' \dontrun{
+#' 
+#' }
+search_for_newer_data = function(archive, ..., 
+                                 source=c("data", path_home("Downloads")), 
+                                 target="data", 
+                                 ask=TRUE, advice=TRUE){
+  check_dots_empty()
+  project_name = parse_file_projname(archive)
+  project_date = parse_file_datetime(archive)
+  
+  source = source[dir_exists(source)]
+  files = dir_ls(source, type="file", fail=FALSE, 
+                 regexp=glue(".*{project_name}.*\\.zip"))
+  if(length(files)==0){
+    cli_inform(c("No project files were found in {.arg source}."))
+    return(invisible(FALSE))
+  }
   
   files_dates = parse_file_datetime(files)
   max_date = max(files_dates, na.rm=TRUE)
-  
-  if(max_date > datetime_extraction){
-    date_diff = round(max_date - datetime_extraction)
-    last_file = files[files_dates==max_date]
-    last_file_short = path_file(last_file) %>% unique()
-    if(length(last_file_short)>1)  rlang::abort("Several last_file_short", .internal=TRUE)
-    if(length(last_file)>1) last_file=last_file[1]  #same file in several paths
-    last_path = path[str_detect(last_file, path)]
-      
-    cli_inform("There is a database in {.path {last_path}} that is {date_diff} days more recent than the current extraction ({date_extraction}).")
+  if(max_date > project_date){
+    date_diff = as.numeric(round(max_date - project_date))
+    newest_file = files[files_dates==max_date]
+    newest_file_name = path_file(newest_file) %>% unique()
+    if(length(newest_file_name)>1)  cli_abort("Several `newest_file_name`", .internal=TRUE)
+    if(length(newest_file)>1) newest_file=newest_file[1]  #same file in several paths
+    last_path = source[str_detect(newest_file, source)]
+    already = if(last_path==target) "already " else ""
+    cli_inform(c("!"="There is {already}a database in {.path {last_path}} that is {col_red(date_diff)} days 
+                 more recent than the current extraction ({as.Date(max_date)} vs {as.Date(project_date)})."))
     
-    from=last_file
-    to=paste0("data/", last_file_short)
+    from = newest_file
+    to = path(target, newest_file_name)
+    
+    if(last_path==target){
+      cli_inform(c("You might want to update your code to import the newer database.", 
+                   i="Newer filename: {.val {newest_file_name}}"))
+      return(invisible(newest_file))
+    }
+    
     user_input = 2
-    if(last_path=="data"){
-      cli_inform("You might want to update your code to import the newer database.")
-    } else if(ask){
-      user_input = cli_menu("Would you like to copy it to {.path /data}?", choices=c("Yes", "No"))
-      if(user_input==1){
-        # ok=file.copy(from, to, copy.date=TRUE)
-        ok=file_copy(from, to)
-        if(!ok) stop("Error during copy")
-      }
+    if(is.numeric(ask)) user_input = ask #for tests
+    else if(ask){
+      user_input = cli_menu("Would you like to copy it to {.path {target}}?", choices=c("Yes", "No"))
     } 
-    if(isTRUE(advice) || is.null(advice) && user_input==2){
-      cli_inform(c("Use the following code to copy it to {.path data}:",
-                   " "='{.emph  file.copy("{from}", \n"{to}", \ncopy.date=TRUE)}'))
+    
+    if(user_input==1){ #1=Yes
+      file_copy(from, to)
+    } else if(isTRUE(advice)){
+      # cli_inform(c("Run the following code to copy it to {.path target}:",
+      #              " "='{.emph file.copy("\n\t{from}", \n"{to}", \ncopy.date=TRUE)}'))
+      # cli_inform(c("Run the following code to copy it to {.path target}:",
+      #              " "='{.emph file.copy(}',
+      #              "  "='{.emph "{from}",}',
+      #              "  "='{.emph "{to}", }',
+      #              "  "='{.emph copy.date=TRUE}',
+      #              " "='{.emph )}'
+      # ))
+      cli_inform("")
+      cli_inform(c("Run the following code to copy it to {.arg target}:"))
+      f = function(x) glue('  "{x}",')
+      cli_code(
+        'file.copy(',
+        f(from),
+        f(to),
+        '  "copy.date=TRUE", ', 
+        ')'
+      )
     }
     
   }
   
-  invisible(last_file)
+  invisible(newest_file)
 }